wip: preserved partial work (auto, session did not succeed)

+++ src/task_state_model.py
@@ -50,7 +50,7 @@ class ExecutionState(AliasModel):
     status: TaskExecutionStatus = "backlog"
     blocked_reason: str | None = Field(default=None, alias="blockedReason")
     retry_count: int = Field(default=0, alias="retryCount")
-    confidence: float | None = Field(default=None, ge=0.0, le=1.0)
+    confidence: int | None = Field(default=None, ge=0, le=1)
     last_session_id: str | None = Field(default=None, alias="lastSessionId")
